autosequences/launch/pyproject.toml
[project]
name = "launch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "synnax>=0.48.0",
    "termcolor>=3.2.0",
    "yaspin>=3.3.0",
]
